// libs/apps/auth/src/lib/components/login/types.ts
import { ReactNode } from 'react';

export interface SavedWallet {
  id: string;
  name: string;
  icon?: string;
}

export interface AccessMethod {
  id: string;
  icon: string;
  text: string;
  disabled?: boolean;
}

export enum ACCESS_METHOD_IDS {
  PRIVATE_KEY = 'PRIVATE_KEY',
  MNEMONIC = 'MNEMONIC',
  KEYSTORE = 'KEYSTORE',
  LEDGER = 'LEDGER'
}